Created new common.py and tf_utils.py.
